Initial commit
I had a six year old repo by the same name. This is entirely new
code.

The previous code was partially moved to FillArrays. I do not appear
as a contributor on that repo, but I actually did some work on
it, in the issues.

FillArrays has `Eye`. But the type here `Id` is more performant,
at least for some things. It is also simpler. It should not be
difficult to make sure that `Id` has better or equal performance
for everything.